Bug fixes 1) pip_wait family functions are rewritten 2) many bugs are fixed now

+++ test/loop.sh
@@ -36,8 +36,10 @@ print_usage()
 finalize()
 {
     if [ x"$TMP" != x ]; then
-	echo "Logfile: $FILE";
-	mv $TMP $FILE;
+	if [ -f $TMP ]; then
+	    echo "Logfile: $FILE";
+	    mv $TMP $FILE;
+	fi
     fi
 }
